Create README.md
This site is meant as a JS resource.  As I continue to learn different tools and libraries, it won't hurt to have a list of JS foundational ideas and where to find more info when needed. And to get the most out of LeetCode I think I need to have a firm grasp of the basics.